refactor: extract analysis request models
